+++ tests/Fixtures/AppBundle/Entity/Topic.php
@@ -1,0 +1,63 @@
+<?php
+/*******************************************************************************
+ *  This file is part of the GraphQL Bundle package.
+ *
+ *  (c) YnloUltratech <support@ynloultratech.com>
+ *
+ *  For the full copyright and license information, please view the LICENSE
+ *  file that was distributed with this source code.
+ ******************************************************************************/
+
+namespace Ynlo\GraphQLBundle\Tests\Fixtures\AppBundle\Entity;
+
+use Doctrine\ORM\Mapping as ORM;
+use Ynlo\GraphQLBundle\Annotation as GraphQL;
+
+/**
+ * @ORM\Entity()
+ *
+ * @GraphQL\ObjectType()
+ * @GraphQL\QueryList()
+ * @GraphQL\MutationAdd()
+ * @GraphQL\MutationUpdate()
+ * @GraphQL\MutationDelete()
+ */
+class Topic
+{
+    /**
+     * @var int|null
+     */
+    protected $id;
+
+    /**
+     * @var string
+     *
+     * @ORM\Column(type="string")
+     */
+    protected $name;
+
+    /**
+     * @var string
+     *
+     * @ORM\OneToMany(targetEntity="Post", mappedBy="topic", fetch="EXTRA_LAZY")
+     */
+    protected $posts;
+
+    /**
+     * User constructor.
+     *
+     * @param int|null $id
+     */
+    public function __construct(?int $id = null)
+    {
+        $this->id = $id;
+    }
+
+    /**
+     * @return int|null
+     */
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+}