+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 v1.1.3 - [Unreleased]
 ----
- * Some minors bugs fixed
+ * Added input type guesser to guess the correctness GraphQL type in forms fields [#10](https://github.com/ynloultratech/graphql-bundle/issues/10)
  * Removed experimental `roles` and role AuthorizationChecker.
+ * Some minors bugs fixed
 
 v1.1.2 - 2018-06-27
 ----